raft: move `poll_output()` from `fsm` to `server`
`server` was the only user of this function and it can now be
implemented using `fsm`'s public interface.

In later commits we'll extend the new member of `server` (currently
called `poll_fsm_output`) to also subscribe to events that `server`
itself generates (not only those generated by `fsm`), so `io_fiber` can
wake up and react to these events.