topology_coordinator: clear obsolete generations earlier
We want to clear CDC generations that are no longer needed (because all
writes are already using a new generation) so they don't take space and
are not sent during snapshot transfers (see e.g.
scylladb/scylladb#17545).

The condition used previously was that we clear generations which were
closed (i.e., a new generation started at this time) more than 24h ago.
This is a safe choice, but too conservative: we could easily end up with
a large number of obsolete generations if we boot multiple nodes during
24h (which is especially easy to do with tablets.)

Change this bound from 24h to `5s + ring_delay`. The choice is explained
in a comment in the code.

Ref: scylladb/scylladb#17545